Proposed change to MAGIC Monero Fund document
Voted for on the regular MAGIC Monero Fund committee meeting of 2023-04-18 by the following individuals in favor:

* artlimber
* Kayaba
* monerobull
* Rucknium

Did not vote (was absent):

* MajesticBank

---

For this to be merged, it needs MAGIC Board approval. Then, the MAGIC Board may at its discretion require a voter vote. Then, the MAGIC Board will announce the change will occur 30 days from approval, or the MAGIC Broad will waive this requirement.